Annotate the 'self' parameter in methods calls so the code is easier to read.
Currently the code is difficult to read in part because there are
multiple classes in a file and the classes have, in many cases,
identical method names...thus complicating knowing where one is
in the code.

This update uses PEP 563 (https://www.python.org/dev/peps/pep-0563),
which begins support with in Python 3.7, and becomes default in
Python 3.10.